ecg_qa_system/requirements.txt
torch>=2.0.1
transformers>=4.30.0
wfdb>=4.0.0
matplotlib>=3.7.1
Pillow>=9.5.0
tqdm>=4.65.0
scipy>=1.10.1